+++ src/main/java/_1danhebojo/coalarm/coalarm_service/domain/alert/controller/response/alertHistory/AlertHistoryResponse.java
@@ -16,7 +16,7 @@ public class AlertHistoryResponse {
 
     public AlertHistoryResponse(AlertHistory alertHistory) {
         this.alertHistoryId = alertHistory.getAlertHistoryId();
-        this.userId = alertHistory.getAlert().getUserId();
+        this.userId = alertHistory.getAlert().getUser().getUserId();
         this.alert = new AlertResponse(alertHistory.getAlert()); // Alert 정보 포함
         this.registeredDate = alertHistory.getRegisteredDate();
     }