+++ src/main/java/_1danhebojo/coalarm/coalarm_service/domain/alert/repository/jpa/AlertJpaRepository.java
@@ -7,6 +7,7 @@ import _1danhebojo.coalarm.coalarm_service.domain.alert.repository.entity.*;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
@@ -22,15 +23,16 @@ public interface AlertJpaRepository extends JpaRepository<Alert, Long> {
             "JOIN FETCH a.coin c " +
             "JOIN FETCH a.user u " +
             "WHERE (:active IS NULL OR a.active = :active) " +
-            "AND (:filter IS NULL OR c.symbol = :filter)")
-    Page<Alert> findAlertsByFilter(@Param("active") Boolean active, @Param("filter") String filter, Pageable pageable);
+            "AND (:filter IS NULL OR c.symbol = :filter)" +
+            "And u.userId = :userId")
+    Page<Alert> findAlertsByFilter(@Param("active") Boolean active, @Param("filter") String filter, Pageable pageable, long userId);
 
     // 사용자가 등록한 모든 활성화된 알람 조회
     @Query("SELECT a " +
             "FROM Alert a " +
             "JOIN FETCH a.coin c " +
             "JOIN FETCH a.user u " +
-            "WHERE a.userId = :userId AND a.active = true")
+            "WHERE a.user.userId = :userId AND a.active = true")
     List<Alert> findActiveAlertsByUserId(Long userId);
 
     @Query("SELECT a " +
@@ -40,6 +42,8 @@ public interface AlertJpaRepository extends JpaRepository<Alert, Long> {
             "WHERE a.active = true")
     List<Alert> findAllActiveAlerts();
 
+    @Modifying
+    @Query("DELETE FROM Alert a WHERE a.user.userId = :userId")
     void deleteAlertByUserId(Long userId);
 
     @Query("SELECT c " +